arruamndo crud com bcrypt para a senha

+++ src/users/users.service.ts
@@ -1,24 +1,45 @@
-import { HttpCode, Injectable } from "@nestjs/common";
+import { HttpCode, HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { User } from "@prisma/client";
 import { UUID } from "crypto";
 import { CreateUserDto } from "src/dto/CreateUserDTO";
 import { PrismaService } from "src/infra/prisma.service";
 import { UpdateUserDTO } from "src/dto/UpdateUserDTO";
+import * as bcrypt from 'bcrypt';
+
 
 @Injectable()
 export class UsersService {
     constructor(private readonly prisma: PrismaService) {}
 
-    async create(data: CreateUserDto): Promise<User> {
-        return this.prisma.user.create({ data });
+    async create(data: CreateUserDto) {
+        const userAlreadyExists = await this.findByEmail(data.email);
+        if (userAlreadyExists) {
+            throw new HttpException('User already exists', HttpStatus.BAD_REQUEST);
+        }else {
+        const user = {
+            ...data,
+            password: await bcrypt.hash(data.password, 10),
+        };
+        const create =  this.prisma.user.create({ data: user });
+
+        return {
+            ...create,
+            password: undefined,
+        };
+    }
+
     }
 
     async findAll(): Promise<User[]> {
         return this.prisma.user.findMany();
     }
 
+    async findByEmail(email: string): Promise<User> {
+        return this.prisma.user.findFirst({ where: { email } });
+    }
+
     async findId(id: UUID): Promise<User> {
-        const user = await this.prisma.user.findFirst({ where: { id } });
+        const user = await this.prisma.user.findUnique({ where: { id } });
         
         return user;
     }
@@ -31,13 +52,20 @@ export class UsersService {
         return this.prisma.user.update({ where: { id }, data });
     }
 
-    async delete(id: UUID): Promise<User> {
-        return this.prisma.user.delete({ where: { id } });
+    async delete(id: UUID) {
+        
+        const user = await this.prisma.user.findFirst({ where: { id } });
+        
+        if (!user) {
+            throw new HttpException('Usuário não encontrado', HttpStatus.NOT_FOUND);
+        } else {
+            
+            await this.prisma.user.delete({ where: { id } });
+            return { message: 'Usuário excluído com sucesso' };
+        }
     }
 
-    async findByEmail(email: string): Promise<User> {
-        return this.prisma.user.findFirst({ where: { email } });
-    }
+    
 }
 
 export default UsersService;
